+++ onnx-tracer/src/graph/node.rs
@@ -466,11 +466,11 @@ impl Node {
         }
     }
 
-    fn imm(&self) -> Option<Tensor<i128>> {
+    fn imm(&self) -> Option<Tensor<i32>> {
         match &self.opkind {
             SupportedOp::Constant(constant) => Some(constant.quantized_values.clone()),
             SupportedOp::Nonlinear(LookupOp::Div { denom }) => {
-                Some(Tensor::from((0..MAX_TENSOR_SIZE).map(|_| denom.0 as i128)))
+                Some(Tensor::from((0..MAX_TENSOR_SIZE).map(|_| denom.0 as i32)))
             }
             _ => None,
         }
@@ -479,11 +479,11 @@ impl Node {
 
 /// A single operation in a [crate::graph::Model].
 
-// TODO: generic quantization
+// TODO(AntoineF4C5): generic quantization
 #[derive(Clone, Debug, PartialEq)]
 pub enum SupportedOp {
     /// A linear operation.
-    Linear(PolyOp<i128>),
+    Linear(PolyOp<i32>),
     /// A nonlinear operation.
     Nonlinear(LookupOp),
     /// A hybrid operation.
@@ -491,7 +491,7 @@ pub enum SupportedOp {
     /// An input node (e.g., model input or placeholder).
     Input(Input),
     /// A constant value node (e.g., weights, biases, or fixed tensors).
-    Constant(Constant<i128>),
+    Constant(Constant<i32>),
     /// An unknown or unsupported operation.
     Unknown(Unknown),
     /// An operation whose inputs have been rescaled for homogeneity.
@@ -515,7 +515,7 @@ impl From<&SupportedOp> for ONNXOpcode {
     }
 }
 
-// TODO: generic quantization
+// TODO(AntoineF4C5): generic quantization
 impl SupportedOp {
     ///
     pub fn is_lookup(&self) -> bool {
@@ -550,7 +550,7 @@ impl SupportedOp {
     }
 
     ///
-    pub fn get_constant(&self) -> Option<&Constant<i128>> {
+    pub fn get_constant(&self) -> Option<&Constant<i32>> {
         match self {
             SupportedOp::Constant(op) => Some(op),
             _ => None,
@@ -558,7 +558,7 @@ impl SupportedOp {
     }
 
     ///
-    pub fn get_mutable_constant(&mut self) -> Option<&mut Constant<i128>> {
+    pub fn get_mutable_constant(&mut self) -> Option<&mut Constant<i32>> {
         match self {
             SupportedOp::Constant(op) => Some(op),
             _ => None,
@@ -569,7 +569,7 @@ impl SupportedOp {
     fn homogenous_rescale(
         &self,
         in_scales: Vec<crate::Scale>,
-    ) -> Result<Box<dyn Op<i128>>, Box<dyn Error>> {
+    ) -> Result<Box<dyn Op<i32>>, Box<dyn Error>> {
         use crate::graph::utilities::homogenize_input_scales;
 
         let inputs_to_scale = self.requires_homogenous_input_scales();
@@ -580,7 +580,7 @@ impl SupportedOp {
 
     /// Since each associated value of `SupportedOp` implements `Op`, let's define a
     /// helper method to retrieve it.
-    fn as_op(&self) -> &dyn Op<i128> {
+    fn as_op(&self) -> &dyn Op<i32> {
         match self {
             SupportedOp::Linear(op) => op,
             SupportedOp::Nonlinear(op) => op,
@@ -597,7 +597,7 @@ impl SupportedOp {
         match self {
             SupportedOp::Input(op) => Node {
                 opkind: self.clone(),
-                out_scale: <Input as Op<i128>>::out_scale(op, vec![]).unwrap(),
+                out_scale: <Input as Op<i32>>::out_scale(op, vec![]).unwrap(),
                 inputs,
                 out_dims,
                 idx,
@@ -621,7 +621,7 @@ impl SupportedOp {
             },
             SupportedOp::Nonlinear(op) => Node {
                 opkind: self.clone(),
-                out_scale: <LookupOp as Op<i128>>::out_scale(op, vec![1]).unwrap(),
+                out_scale: <LookupOp as Op<i32>>::out_scale(op, vec![1]).unwrap(),
                 inputs,
                 out_dims,
                 idx,
@@ -629,7 +629,7 @@ impl SupportedOp {
             },
             SupportedOp::Hybrid(op) => Node {
                 opkind: self.clone(),
-                out_scale: <HybridOp as Op<i128>>::out_scale(op, vec![1]).unwrap(),
+                out_scale: <HybridOp as Op<i32>>::out_scale(op, vec![1]).unwrap(),
                 inputs,
                 out_dims,
                 idx,
@@ -645,7 +645,7 @@ impl SupportedOp {
             },
             SupportedOp::Rescaled(op) => Node {
                 opkind: self.clone(),
-                out_scale: <Rescaled as Op<i128>>::out_scale(op, vec![1]).unwrap(),
+                out_scale: <Rescaled as Op<i32>>::out_scale(op, vec![1]).unwrap(),
                 inputs,
                 out_dims,
                 idx,
@@ -653,7 +653,7 @@ impl SupportedOp {
             },
             SupportedOp::RebaseScale(op) => Node {
                 opkind: self.clone(),
-                out_scale: <RebaseScale as Op<i128>>::out_scale(op, vec![1]).unwrap(),
+                out_scale: <RebaseScale as Op<i32>>::out_scale(op, vec![1]).unwrap(),
                 inputs,
                 out_dims,
                 idx,
@@ -663,10 +663,10 @@ impl SupportedOp {
     }
 }
 
-// TODO: generic quantization
-impl From<Box<dyn Op<i128>>> for SupportedOp {
-    fn from(value: Box<dyn Op<i128>>) -> Self {
-        if let Some(op) = value.as_any().downcast_ref::<PolyOp<i128>>() {
+// TODO(AntoineF4C5): generic quantization
+impl From<Box<dyn Op<i32>>> for SupportedOp {
+    fn from(value: Box<dyn Op<i32>>) -> Self {
+        if let Some(op) = value.as_any().downcast_ref::<PolyOp<i32>>() {
             return SupportedOp::Linear(op.clone());
         };
 
@@ -682,7 +682,7 @@ impl From<Box<dyn Op<i128>>> for SupportedOp {
             return SupportedOp::Input(op.clone());
         };
 
-        if let Some(op) = value.as_any().downcast_ref::<Constant<i128>>() {
+        if let Some(op) = value.as_any().downcast_ref::<Constant<i32>>() {
             return SupportedOp::Constant(op.clone());
         };
 
@@ -702,12 +702,9 @@ impl From<Box<dyn Op<i128>>> for SupportedOp {
     }
 }
 
-// TODO: generic quantization
-impl Op<i128> for SupportedOp {
-    fn f(
-        &self,
-        inputs: &[Tensor<i128>],
-    ) -> Result<ForwardResult<i128>, crate::tensor::TensorError> {
+// TODO(AntoineF4C5): generic quantization
+impl Op<i32> for SupportedOp {
+    fn f(&self, inputs: &[Tensor<i32>]) -> Result<ForwardResult<i32>, crate::tensor::TensorError> {
         self.as_op().f(inputs)
     }
 
@@ -723,7 +720,7 @@ impl Op<i128> for SupportedOp {
         self.as_op().requires_homogenous_input_scales()
     }
 
-    fn clone_dyn(&self) -> Box<dyn Op<i128>> {
+    fn clone_dyn(&self) -> Box<dyn Op<i32>> {
         self.as_op().clone_dyn()
     }
 
@@ -753,23 +750,23 @@ pub struct Rescaled {
     pub scale: Vec<(usize, u128)>,
 }
 
-// TODO: generic quantization
-impl Op<i128> for Rescaled {
+// TODO(AntoineF4C5): generic quantization
+impl Op<i32> for Rescaled {
     fn as_any(&self) -> &dyn std::any::Any {
         self
     }
-    fn f(&self, x: &[Tensor<i128>]) -> Result<ForwardResult<i128>, TensorError> {
+    fn f(&self, x: &[Tensor<i32>]) -> Result<ForwardResult<i32>, TensorError> {
         if self.scale.len() != x.len() {
             return Err(TensorError::DimMismatch("rescaled inputs".to_string()));
         }
         let mut rescaled_inputs = vec![];
         let inputs = &mut x.to_vec();
         for (i, ri) in inputs.iter_mut().enumerate() {
-            let mult_tensor = Tensor::from([self.scale[i].1 as i128].into_iter());
+            let mult_tensor = Tensor::from([self.scale[i].1 as i32].into_iter());
             let res = (ri.clone() * mult_tensor)?;
             rescaled_inputs.push(res);
         }
-        Op::<i128>::f(&*self.inner, &rescaled_inputs)
+        Op::<i32>::f(&*self.inner, &rescaled_inputs)
     }
 
     fn as_string(&self) -> String {
@@ -783,10 +780,10 @@ impl Op<i128> for Rescaled {
             .map(|(a, b)| a + multiplier_to_scale(b.1 as f64))
             .collect();
 
-        Op::<i128>::out_scale(&*self.inner, in_scales)
+        Op::<i32>::out_scale(&*self.inner, in_scales)
     }
 
-    fn clone_dyn(&self) -> Box<dyn Op<i128>> {
+    fn clone_dyn(&self) -> Box<dyn Op<i32>> {
         Box::new(self.clone()) // Forward to the derive(Clone) impl
     }
 }
@@ -867,13 +864,13 @@ impl RebaseScale {
     }
 }
 
-// TODO: generic quantization
-impl Op<i128> for RebaseScale {
+// TODO(AntoineF4C5): generic quantization
+impl Op<i32> for RebaseScale {
     fn as_any(&self) -> &dyn std::any::Any {
         self
     }
-    fn f(&self, x: &[Tensor<i128>]) -> Result<ForwardResult<i128>, TensorError> {
-        let mut res = Op::<i128>::f(&*self.inner, x)?;
+    fn f(&self, x: &[Tensor<i32>]) -> Result<ForwardResult<i32>, TensorError> {
+        let mut res = Op::<i32>::f(&*self.inner, x)?;
         let ri = res.output;
         let rescaled = crate::tensor::ops::nonlinearities::const_div(&ri, self.multiplier);
         res.output = rescaled;
@@ -903,7 +900,7 @@ impl Op<i128> for RebaseScale {
         lookups
     }
 
-    fn clone_dyn(&self) -> Box<dyn Op<i128>> {
+    fn clone_dyn(&self) -> Box<dyn Op<i32>> {
         Box::new(self.clone()) // Forward to the derive(Clone) impl
     }
 }
@@ -944,7 +941,7 @@ impl Tabled for Node {
             self.opkind
                 .required_lookups()
                 .iter()
-                .map(<LookupOp as Op<i128>>::as_string)
+                .map(<LookupOp as Op<i32>>::as_string)
                 .collect_vec()
         )));
         fields
@@ -992,9 +989,9 @@ impl PartialEq for Node {
 /// rescale_const_with_single_use(constant, input_scales, constant_node.num_uses())?;
 /// ```
 
-// TODO: generic quantization
+// TODO(AntoineF4C5): generic quantization
 fn rescale_const_with_single_use(
-    constant: &mut Constant<i128>,
+    constant: &mut Constant<i32>,
     in_scales: Vec<crate::Scale>,
     num_uses: usize,
 ) -> Result<(), Box<dyn Error>> {
